+++ bin/workflow_glue/report.py
@@ -281,7 +281,7 @@ def main(args):
             between the supported kits. They are:
             """
         )
-        p("3prime and multiome kits:")
+        p("3prime, multiome and visium kits:")
         ul(
             li("Adapter1: Read1"),
             li("Adapter2: TSO")